tokenを使った、編集、削除制限機能実装

+++ controllers/feed.js
@@ -128,6 +128,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authenticated");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -158,12 +163,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authenticated");
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: "Post delete successfully" });
     })
     .catch((err) => {
